fix package directory generation for m-types used as wrappers inside opaques
This covers the following case:

 - type X is an opaque, converted to type A
 - to marshall type A, we manually define type B
 - B contains m-version of opaque types (e.g. uses base::Pose_m)

The current code would fail at the generation as B is NOT used
as an intermediate for an opaque, but, regardless, is a m-type
regardless from the p.o.v. of the type system. It would not be
registered in the package directory and we would get an infinite
loop in #link_to

Fix by generating a page for these types.