deb: adapt all include path to using -isystem instead of -I
When mixing debian and local packages, the order of include directories matters,
but the include directories from the debian package need to be +always+ of lower
priority.
I considered two options here and taking the simpler one (b):
(a) change either the include dir install path into an isolated folder and provide this in the pc file
or (b) change from -I to -isystem.